swap out accession formats
