+++ src/encoded/schemas/antibody_lot.json
@@ -5,37 +5,22 @@
     "$schema": "http://json-schema.org/draft-04/schema#",
     "type": "object",
     "required": ["product_id", "source", "host_organism"],
-    "identifyingProperties": ["uuid","accession","aliases"],
+    "identifyingProperties": ["uuid","aliases"],
     "additionalProperties": false,
     "mixinProperties": [
         {"$ref": "mixins.json#/aliases" },
         {"$ref": "mixins.json#/schema_version" },
         {"$ref": "mixins.json#/uuid" },
-        {"$ref": "mixins.json#/accession" },
-        {"$ref": "mixins.json#/accessioned_status" },
+        {"$ref": "mixins.json#/standard_status" },
         {"$ref": "mixins.json#/submitted" },
         {"$ref": "mixins.json#/source" },
         {"$ref": "mixins.json#/url" },
         {"$ref": "mixins.json#/notes" },
         {"$ref": "mixins.json#/dbxrefs"}
-
     ],
     "properties": {
         "schema_version": {
             "default": "1"
-        },
-        "accession": {
-            "accessionType": "AB"
-        },
-        "alternate_accessions": {
-            "items": {
-                "title": "Alternate Accession",
-                "description": "An accession previously assigned to an object that has been merged with this object.",
-                "comment": "Only accessions of objects that have status equal replaced will work here.",
-                "type": "string",
-                "format": "accession",
-                "pattern": "^(ENCAB\\d{3}[A-Z]{3})$|^(TSTAB\\d{6})$"
-            }
         },
         "lot_id_alias": {
             "title": "Lot ID aliases",
