+++ src/encoded/schemas/biosample.json
@@ -23,7 +23,7 @@
                 "comment": "Only accessions of objects that have status equal replaced will work here.",
                 "type": "string",
                 "format": "accession",
-                "pattern": "^(ENCBS\\d{3}[A-Z]{3})$|^(TSTBS\\d{6})$"
+                "pattern": "^(LATBS\\d{3}[A-Z]{3})$"
             }
         },
         "biosample_ontology": {
