+++ src/encoded/schemas/mixins.json
@@ -55,7 +55,7 @@
     "accession": {
         "accession": {
             "title": "Accession",
-            "description": "A unique identifier to be used to reference the object prefixed with ENC.",
+            "description": "A unique identifier to be used to reference the object prefixed with LAT.",
             "comment": "Do not submit. The accession is assigned by the server.",
             "type": "string",
             "format": "accession",
