+++ src/encoded/schemas/mouse_donor.json
@@ -13,6 +13,19 @@
         {"$ref": "donor.json#/properties"}
     ],
     "properties": {
+        "accession": {
+            "accessionType": "MD"
+        },
+        "alternate_accessions": {
+            "items": {
+                "title": "Alternate accession",
+                "description": "An accession previously assigned to an object that has been merged with this object.",
+                "comment": "Only accessions of objects that have status equal replaced will work here.",
+                "type": "string",
+                "format": "accession",
+                "pattern": "^(LATMD\\d{3}[A-Z]{3})$"
+            }
+        },
         "dbxrefs": {
             "items": {
                 "title": "Database external identifier",
