+++ src/encoded/schemas/reference_file_set.json
@@ -11,7 +11,6 @@
         { "$ref": "mixins.json#/aliases"},
         { "$ref": "mixins.json#/schema_version"},
         { "$ref": "mixins.json#/uuid"},
-        { "$ref": "mixins.json#/accession"},
         { "$ref": "mixins.json#/submitted"},
         { "$ref": "mixins.json#/notes"},
         { "$ref": "mixins.json#/description" },
@@ -22,19 +21,6 @@
     "properties": {
         "schema_version": {
             "default": "1"
-        },
-        "accession": {
-            "accessionType": "SR"
-        },
-        "alternate_accessions": {
-            "items": {
-                "title": "Alternate Accession",
-                "description": "An accession previously assigned to an object that has been merged with this object.",
-                "comment": "Only accessions of objects that have status equal replaced will work here.",
-                "type": "string",
-                "format": "accession",
-                "pattern": "^(ENCSR\\d{3}[A-Z]{3})$|^(TSTSR\\d{6})$"
-            }
         },
         "description": {
             "elasticsearch_mapping_index_type": {
