+++ src/encoded/tests/data/inserts/mouse_postnatal_donor.json
@@ -1,6 +1,6 @@
 [
     {
-        "accession": "ENCDO060DDD",
+        "accession": "LATMD060DDD",
         "aliases": [
             "anna-greka:adult_mouse_donorE"
         ],
@@ -23,7 +23,7 @@
         ]
     },
         {
-        "accession": "ENCDO065DDD",
+        "accession": "LATMD065DDD",
         "aliases": [
             "anna-greka:adult_mouse_donorF"
         ],
