+++ src/encoded/tests/data/inserts/mouse_prenatal_donor.json
@@ -1,6 +1,6 @@
 [
     {
-        "accession": "ENCDO050DDD",
+        "accession": "LATMD050DDD",
         "aliases": [
             "anna-greka:embryonic_mouse_donorG"
         ],
@@ -23,7 +23,7 @@
         ]
     },
     {
-        "accession": "ENCDO055DDD",
+        "accession": "LATMD055DDD",
         "aliases": [
             "anna-greka:embryonic_mouse_donorH"
         ],
