+++ src/encoded/tests/data/inserts/organoid.json
@@ -4,7 +4,7 @@
             "anna-greka:organoid_donorC1"
         ],
         "status": "released",
-        "accession": "ENCBS400DDD",
+        "accession": "LATBS400DDD",
         "uuid": "840d1bba-0005-4935-901a-41bc754103b4",
         "description": "organoid from cell culture of prenatal donorC",
         "derived_from": [
@@ -27,7 +27,7 @@
             "anna-greka:organoid_donorC2"
         ],
         "status": "released",
-        "accession": "ENCBS405DDD",
+        "accession": "LATBS405DDD",
         "uuid": "5bda2505-a8bd-40b9-bee2-d56eed19c7b8",
         "description": "organoid from cell culture of prenatal donorC",
         "derived_from": [
@@ -50,7 +50,7 @@
             "anna-greka:organoid_donorD1"
         ],
         "status": "released",
-        "accession": "ENCBS410DDD",
+        "accession": "LATBS410DDD",
         "uuid": "57672176-ce03-45b3-a3b5-2cc0bc477988",
         "description": "organoid from cell culture of prenatal donorD",
         "derived_from": [
@@ -73,7 +73,7 @@
             "anna-greka:organoid_donorD2"
         ],
         "status": "released",
-        "accession": "ENCBS415DDD",
+        "accession": "LATBS415DDD",
         "uuid": "ada165d6-05df-482f-98fb-ababb55ed31b",
         "description": "organoid from cell culture of prenatal donorD",
         "derived_from": [
@@ -96,7 +96,7 @@
             "jimmie-ye:prenatal_organoid1"
         ],
         "status": "in progress",
-        "accession": "ENCBS100SCS",
+        "accession": "LATBS100SCS",
         "uuid": "164042c5-ce91-4ca6-8a96-2d8ea58143c8",
         "description": "prenatal cell organoid 1",
         "derived_from": [
@@ -112,7 +112,7 @@
             "jimmie-ye:prenatal_organoid2"
         ],
         "status": "in progress",
-        "accession": "ENCBS101SCS",
+        "accession": "LATBS101SCS",
         "uuid": "7a0c1810-4d95-430b-9916-ab640a3cfe5d",
         "description": "prenatal cell organoid 2",
         "derived_from": [
@@ -128,7 +128,7 @@
             "jimmie-ye:prenatal_organoid3"
         ],
         "status": "in progress",
-        "accession": "ENCBS102SCS",
+        "accession": "LATBS102SCS",
         "uuid": "20e7a828-f0a5-4f3c-a122-f6385c98a1d5",
         "description": "prenatal cell organoid 3",
         "derived_from": [
