+++ src/encoded/tests/data/inserts/reference_file_set.json
@@ -1,14 +1,18 @@
 [
     {
-        "accession": "ENCSR500ELM",
         "description": "Human: Elements DNA sequences",
+        "aliases": [
+            "lattice:human-references"
+        ],
         "status": "released",
         "submitted_by": "massa.porta@varius.mauris",
         "uuid": "84c2c840-4146-4c93-90cf-67a11a4fc108"
     },
     {
-    	"accession": "ENCSR800RRR",
         "description": "Mouse: Elements DNA sequences",
+        "aliases": [
+            "lattice:mouse-references"
+        ],
         "status": "released",
         "submitted_by": "quisque.consequat@venenatis.magna",
         "uuid": "b556ddc5-9ad1-4cf2-aad2-6c2116e7e7b4"
